+++ service-modules/agent/src/main/java/com/zw/agent/entity/SysPermissionEntity.java
@@ -55,7 +55,7 @@ public class SysPermissionEntity extends BaseEntity {
      * 类型：1-菜单，2-按钮/接口
      */
     @TableField("type")
-    private Byte type;
+    private Integer type;
 
     /**
      * 排序号
@@ -67,5 +67,5 @@ public class SysPermissionEntity extends BaseEntity {
      * 状态：0-禁用，1-启用
      */
     @TableField("status")
-    private Byte status;
+    private Integer status;
 }
